Moved widget files to their own folder
