+++ src/widgets/webcam_widget.py
